CMM-1142 stats create todays card rust (#22502)
* Adding card and basic info

* Add horuly granularity

* Some styling

* Following iOS distribution

* Adding yesterday's data

* Gradiente and style

* Line fixes

* detekt

* Adding tests

* trailing line

* Extractign strings

* Adding pull to refresh

* Using rust library

* Fixing date period

* Updating version

* Calling daily aggregates

* Adding tests

* detekt

* Rename and tests

* Using Long instead of Int for stats

* Other PR suggestions

* Using RS trunk version

+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/todaysstat/TodaysStatsCard.kt
@@ -376,10 +376,10 @@ private fun StatsChart(chartData: ChartData) {
 
 @Composable
 private fun MetricsRow(
-    views: Int,
-    visitors: Int,
-    likes: Int,
-    comments: Int
+    views: Long,
+    visitors: Long,
+    likes: Long,
+    comments: Long
 ) {
     Row(
         modifier = Modifier.fillMaxWidth(),
@@ -457,7 +457,7 @@ private fun SecondaryMetricItem(
     }
 }
 
-private fun formatStatValue(value: Int): String {
+private fun formatStatValue(value: Long): String {
     return when {
         value >= MILLION -> String.format(Locale.getDefault(), "%.1fM", value / MILLION.toDouble())
         value >= THOUSAND -> String.format(Locale.getDefault(), "%.1fK", value / THOUSAND.toDouble())
